[mod] スマホビュー対応(#65)
