+++ view/flock_view/components/UserInfo.vue
@@ -1,6 +1,6 @@
 <template>
   <div>
-    <v-card flat class="pa-3" color="green lighten-4" rounded="lg">
+    <v-card flat class="pa-3" color="green lighten-4" rounded="lg" width="100%" max-width="100%">
       <b><nuxt-link :to="{ name: 'users-id', params: { id: user.id } }">{{ user.name }}</nuxt-link> さんの投稿</b>
       <p class="my-2">{{ user.age }}歳 {{ sexList[user.sex - 1] }}</p>
       <p class="my-2">{{ user.profile }}</p>
